requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-docx==1.1.2
pydantic==2.7.4
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
rapidfuzz==3.9.6
aiofiles==23.2.1
lxml==5.2.2
pytest==8.2.0
